fix(schema-compiler): Missing pre-aggregation filter for partitioned pre-aggregation via view, fix #6623 (#7454)
There was an issue when partitioned pre-aggregation was queried via view. Schema compiler could not find a time dimension to get a range of used pre-aggregation from the query.

As a result, it started to union all partitions. See real example from the issue: https://github.com/cube-js/cube/issues/6623