feat(cubestore): Metastore - increase parallelism to 2 (#8807)
parallelism: The number of threads RocksDB will use for all background operations. Even if you increase the max_background_jobs but keep parallelism as 1, it won’t lead to a lot of performance increase. Ideally, this option should be kept equal to the number of cores in your machine.

max_background_jobs is 2, but parallelism was 1. Must be aligned.