fix(schema-compiler): Collect join hints from subquery join conditions (#9554)
This is necessary to support queries where output references only join subqueries.
Before, when query does have subqeury joins, but does not have any measures/dimensions/... join hints would be empty, and join tree would be null, which is almost okay, but to generate query with join subqueries we need to build join query, so join tree is more-or-less required.