fix(query-orchestrator): Reduce number of refresh key queries (#9809)
Cube generates a burst load of queries for refresh keys. We don't use a queue for cube store queries, but we use an in-memory cache inside QueryQueue. A cache miss leads to a problem where we start executing queries without debouncing.

Cube executes refresh keys in parallel, which DDoSes Cube Store. It's a quick fix to fix this behavior.