refactor(schema-compiler): Drop support for disabling native workers … (#10196)
This flag is enabled by default for a long time and there are not side effect. I am working on speeding up Cube startup and I've figure out that we load babel, which we don't need in the Cube Cloud Runtime, because it uses native transpilers. So, let's remove this env variable and speedup cloud as side effect.

---------

Co-authored-by: Konstantin Burkalev <KSDaemon@gmail.com>